Break up run.rs into four files
