+++ crates/eframe/src/native/mod.rs
@@ -5,3 +5,11 @@ pub mod run;
 /// File storage which can be used by native backends.
 #[cfg(feature = "persistence")]
 pub mod file_storage;
+
+pub(crate) mod winit_integration;
+
+#[cfg(feature = "glow")]
+mod glow_integration;
+
+#[cfg(feature = "wgpu")]
+mod wgpu_integration;
